fix(p1am): coerce _temp_from_tags return to float for CI mypy
CI's quality-gate runs mypy with --follow-imports=skip, so the imported
TemperatureConfig is treated as Any and the deg-C scaling expression returned
Any from a function declared -> float (no-any-return). Wrap the result in
float() so the contract holds under skipped-import type checking.

+++ src/p1am_control_system/backend/temperature_integration.py
@@ -62,7 +62,7 @@ class TemperatureService:
             return 0.0
         cfg = self.controller.config
         temp_pct = float(tags.get(cfg.temp_tag, 0.0))
-        return temp_pct * cfg.temp_full_scale_c / 100.0
+        return float(temp_pct * cfg.temp_full_scale_c / 100.0)
 
     async def _write_relay(self, on: bool) -> bool:
         """Command the heater relay via the client's public coil seam."""
